[naga wgsl-in] Relocate `automatic_conversion_consensus`.
Move `naga::front::wgsl::lower::construction::automatic_conversion_consensus`
into `conversion.rs`, and make it a method of `ExpressionContext`.
